# aria-frontend/openapi/aria.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicSaveDraftRequest": {
        "properties": {
          "campaign_tag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Campaign Tag"
          },
          "company_id": {
            "title": "Company Id",
            "type": "string"
          },
          "content": {
            "maxLength": 5000,
            "minLength": 3,
            "title": "Content",
            "type": "string"
          },
          "cta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cta"
          },
          "intent": {
            "default": "engage",
            "title": "Intent",
            "type": "string"
          },
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "tone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tone"
          },
          "topic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topic"
          }
        },
        "required": [
          "company_id",
          "platform",
          "content"
        ],
        "title": "PublicSaveDraftRequest",
        "type": "object"
      },
...
      "PublicScheduleRequest": {
        "properties": {
          "approval_mode": {
            "default": "human",
            "title": "Approval Mode",
            "type": "string"
          },
          "company_id": {
            "title": "Company Id",
            "type": "string"
          },
          "manual_override": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Override"
          },
          "post_id": {
            "title": "Post Id",
            "type": "string"
          },
          "targets": {
            "items": {
              "$ref": "#/components/schemas/PublicScheduleTarget"
            },
            "minItems": 1,
            "title": "Targets",
            "type": "array"
          }
        },
        "required": [
          "post_id",
          "company_id",
          "targets"
        ],
        "title": "PublicScheduleRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ARIA LLM Orchestration Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/companies/{company_id}/posts": {
      "get": {
        "operationId": "public_list_company_posts_v1_companies__company_id__posts_get",
        "parameters": [
          {
            "in": "path",
            "name": "company_id",
            "required": true,
            "schema": {
              "title": "Company Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Public List Company Posts V1 Companies  Company Id  Posts Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Public List Company Posts",
        "tags": [
          "public-runtime"
        ]
      }
    },
...
    "/v1/posts/drafts": {
      "post": {
        "operationId": "public_save_draft_v1_posts_drafts_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicSaveDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Public Save Draft V1 Posts Drafts Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Public Save Draft",
        "tags": [
          "public-runtime"
        ]
      }
    },
    "/v1/posts/generate": {
      "post": {
        "operationId": "public_generate_post_v1_posts_generate_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicGeneratePostRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Public Generate Post V1 Posts Generate Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Public Generate Post",
        "tags": [
          "public-runtime"
        ]
      }
    },
    "/v1/posts/{post_id}": {
      "get": {
        "operationId": "public_get_post_v1_posts__post_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "post_id",
            "required": true,
            "schema": {
              "title": "Post Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Public Get Post V1 Posts  Post Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Public Get Post",
        "tags": [
          "public-runtime"
        ]
      }
    },
    "/v1/schedules": {
      "post": {
        "operationId": "public_create_schedule_v1_schedules_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicScheduleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Public Create Schedule V1 Schedules Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Public Create Schedule",
        "tags": [
          "public-runtime"
        ]
      }
    },
    "/v1/schedules/{schedule_id}": {
      "get": {
        "operationId": "public_get_schedule_v1_schedules__schedule_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "schedule_id",
            "required": true,
            "schema": {
              "title": "Schedule Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Public Get Schedule V1 Schedules  Schedule Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Public Get Schedule",
        "tags": [
          "public-runtime"
        ]
      }
    },
...
  }
}